adicao de testes xunit

+++ APICalculaJuros/Domains/TaxaJuros.cs
@@ -2,6 +2,6 @@ namespace APICalculaJuros.Domains
 {
     public class TaxaJuros
     {
-        public float taxaJuros { get; set; }
+        public double taxaJuros { get; set; }
     }
 }
